Clean up scrollable.dart for 2D
