+++ packages/flutter/lib/src/widgets/reorderable_list.dart
@@ -16,6 +16,7 @@ import 'scroll_controller.dart';
 import 'scroll_physics.dart';
 import 'scroll_view.dart';
 import 'scrollable.dart';
+import 'scrollable_helpers.dart';
 import 'sliver.dart';
 import 'sliver_prototype_extent_list.dart';
 import 'ticker_provider.dart';
