+++ packages/flutter/lib/widgets.dart
@@ -115,6 +115,7 @@ export 'src/widgets/scroll_position_with_single_context.dart';
 export 'src/widgets/scroll_simulation.dart';
 export 'src/widgets/scroll_view.dart';
 export 'src/widgets/scrollable.dart';
+export 'src/widgets/scrollable_helpers.dart';
 export 'src/widgets/scrollbar.dart';
 export 'src/widgets/selectable_region.dart';
 export 'src/widgets/selection_container.dart';
